+++ apps/www/src/app/(main)/(docs)/components/autocomplete/examples/error-state.tsx
@@ -59,7 +59,7 @@ export const AutocompleteErrorState = () => {
     <div>
       <Autocomplete
         label="Animal"
-        helperText="Please choose one of the options"
+        errorMessage="Please choose one of the options"
         isInvalid
       >
         {animals.map((animal) => (