+++ apps/www/src/app/(main)/(docs)/components/input/examples/error-state.tsx
@@ -4,7 +4,7 @@ export const InputErrorState = () => {
   return (
     <Input
       label="Username"
-      helperText="Please use a unique username"
+      errorMessage="Please use a unique username"
       isInvalid
     />
   );