+++ apps/www/src/app/(main)/(docs)/components/input/examples/usage.tsx
@@ -2,10 +2,6 @@ import { Input } from "@jamsr-ui/react";
 
 export const InputUsage = () => {
   return (
-    <Input
-      className="top-[-0.6px]"
-      label="Enter your email"
-      type="email"
-    />
+    <Input className="top-[-0.6px]" label="Enter your email" type="email" />
   );
 };