+++ packages/components/rhf/src/select.tsx
@@ -31,7 +31,7 @@ export const RHFSelect = <T extends FieldValues>(props: Props<T>) => {
             onBlur={onBlur}
             isInvalid={!!error}
             data-invalid={!!error}
-            helperText={error?.message}
+            errorMessage={error?.message}
             isMultiple={isMultiple}
             {...restProps}
           />