+++ packages/components/rhf/src/tags-input.tsx
@@ -29,7 +29,7 @@ export const RHFTagsInput = <T extends FieldValues>(props: Props<T>) => {
             onBlur={onBlur}
             isInvalid={!!error}
             data-invalid={!!error}
-            helperText={error?.message}
+            errorMessage={error?.message}
             {...restProps}
           />
         );