+++ packages/components/textarea/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/input
 
+## 0.1.5
+
+### Patch Changes
+
+- feat: separate errormessage prop
+- Updated dependencies
+  - @jamsr-ui/input@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes